feat: basic AP <-> lysand conversion (#4)
* feat: Update API domain variable name

* save changes, fake commit

* feat: function to receive lysand note

* [feat]: lysand to ap for posts and users

* feat: Add .env file to gitignore and update dependencies

The commit adds the `.env` file to the `.gitignore` and updates the dependencies in the `Cargo.toml` and `Cargo.lock` files. This change ensures that sensitive environment variables are not committed to the repository and keeps the dependencies up to date.

* feat: Add db_post_from_url function

The commit adds the `db_post_from_url` function to the `conversion.rs` file. This function retrieves a post from the database based on a given URL. If the post is not found in the database, it fetches the post from the URL, saves it to the database, and returns the post. This change enhances the functionality of the codebase by providing a convenient way to retrieve and store posts.

* feat: Update dependencies and add async-recursion crate

* fix: Refactor lysand note handling in conversion.rs

The commit refactors the lysand note handling in the `conversion.rs` file. It updates the `receive_lysand_note` function to properly handle quoting and replying to notes. This change improves the functionality and readability of the codebase.

* fix: use example as default user

* me oopid

* fix: Refactor lysand note handling in conversion.rs

* fix: fix post printing

* fix: Refactor lysand note handling in conversion.rs

* fix: Refactor lysand note handling in conversion.rs

* fix: make nix-bootstrap executable

* fix: remove unused linux only import

* fix: make person stop screaming at me :(

* feat: remove test code

* fix: update deps

* fix: rm build.rs fully

* feat: standard user to apservice

* fix: format files